+++ docs/ReleaseNotes.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Makes `StreamEntry` constructor public for better unit test experience (#1923 via WeihanLi)
+
 ## 2.2.88
 
 - Connection backoff default is now exponential instead of linear (#1896 via lolodi)